+++ test/couchdb_test.dart
@@ -1,12 +1,7 @@
-import 'dart:io';
-import 'dart:convert';
-
-import 'package:http/http.dart';
-
 import 'package:couchdb/couchdb.dart';
 
 Future<void> main() async {
-  final c = CouchDbClient(username: 'admin', password: 'pass');
+  final c = CouchDbClient(username: 'admin', password: 'pass', auth: 'cookie');
   final da = DatabaseModel(c);
   final ddm = DesignDocumentModel(c);
   final dm = DocumentModel(c);
@@ -16,9 +11,12 @@ Future<void> main() async {
 
   try {
     final headers = <String, String>{'Accept': 'text/plain'};
-    final o = await da.changesIn('denta', feed: 'eventsource');
     
+    final r = await c.authenticate();
+    
+    final o = await da.changesIn('denta');
     await o.forEach(print);
+    print(await c.userInfo());
   } on CouchDbException catch (e) {
     print(e);
   }